Separacion de funciones en multiples archivos .h
