+++ Aprendiendo-C.code-workspace
@@ -1,0 +1,12 @@
+{
+	"folders": [
+		{
+			"path": "."
+		}
+	],
+	"settings": {
+		"files.associations": {
+			"entrada.h": "c"
+		}
+	}
+}
